Shuffling around tile seeders Added the generic color door and key
